Add the Haskell side of the remote runner
This function (`runRemote`) takes channels for inputs and outputs (as well as
logging) and manages an SSH connection to a remote host to drive the remote-runner
and parse its results.  Communication is over the stdin and stdout of the
remote-runner ssh connection.  There are a few threads running (via the async
library) to handle the communication channels.  It tries to tear down in a
reasonable way.