Rewrite the remote runner
It now uses ptrace instead of the context handling support from signals and
setcontext/getcontext/makecontext.

There are two important limitations of the context infrastructure:

1) Not all vector registers are accessible in the ucontext_t; this is especially
   problematic on x86_64.

2) The implementation of setcontext requires at least one register to jump
   through to execute a context.  This means that there will always be one
   register that we can't use in the environment.  On PPC, that was the CTR
   register, which was used to hold the address of the jump target.

With ptrace, neither is a concern (though the ZMM registers don't seem to be
accessible on x86_64 right now, ptrace is more likely to gain access quickly
than getcontext is).

This will also be more robust to strange errors, potentially, since the monitor
process is completely external.

The big annoyance is that the format of registers is different.  We now have an
intermediate format for register states that differs by architecture.  Before we
were just using mcontext_t.