Add a generic helper for unfolding values of kind * into values with a type-level shape parameter
See the Data.Parameterized.Unfold module.  This is an experiment.  It would
allow us to write multiple "shape unfolds" with a single class constraint (to
avoid a proliferation of class constraints for each operation we need to do like
this).  At a high level, it basically exposes the deconstruction of the type
level list at the value level through proxies, only requiring the class
instances to be defined once.

We might be able to generalize this more.

The downside is that it has a tiny bit of boilerplate.