Set up the IORelation learning to actually be runnable
The interface now takes some input configuration to tell it where to find and
store learned IO relations.  It also forks off a number of worker threads and
sets up the necessary solver state.

There was a bit of a refactoring of the remote interface.  The class constraint
for `MachineState` was problematic, as we couldn't reference the `t` type
variable (it is a state thread variable that isn't visible at the entry point to
the learning procedure, since we haven't allocated a nonce generator yet).  The
change makes `MachineState` plain data that is a bit easier to manage.