Refactor test runner infrastructure.
The main change is that `TestCase` now contains a program of type
`[i]`, for `i` an instruction type, instead of a bytestring
corresponding to the assembly of a program. This will allow me to
implement a test runner for the toy arch without bothering with
serialization/deserialization. The reason that it's a list of
instructions, and not a single instruction, is that MCMC synth tests
programs comprising sequences of instructions.