Bug fixes to the remote runner and IORelation learning
The remote runner had a serious memory consistency bug.  The fix is to flush the
icache after writing a new test program to memory.  The problem showed up as
nondeterministic test failures in the IORelation learning.

There are also a few fixes to the logic in the IORelation learning.  In
particular, explicit IORelation learning no longer generates test instructions
with repeated register operands.  The detection logic for implicit operands was
also rewritten to actually seem plausible.