Fix loading formulas from disk.
It might be simpler to just read all formula files in some specified
directories, but apparently we need to know some class instances for
an opcode in order to read its formula from disk. So, we give a list
of class instances for the opcodes and then check if any of them are
available on disk.

I also added some debug printfs to see that things are working.