Instantiate `mcmcSynthesizeOne` at Toy arch.
This works for learning implementations of addition using subtraction
and negation on the toy arch, but the evolution of weights is
suspicious: it should be `num-tests * wrong-out-location-penalty`, but
it converges to about 5 times that, and then drops to zero when a
correct candidate is found.

This code is littered with debug printfs, and other cruft, which I can
clean up once I'm convinced everything is really working.