Add support for formula extraction
This support lives in a new module called `SemMC.Stochastic.Extract`.  One of
the major changes alongside this code is some support metadata to allow us to
learn formulas for opcodes that include immediates.

The original stochastic synthesis paper doesn't actually handle this very well -
in cases where they do learn semantics for instructions with immediates, they
learn a formula for each possible value of the immediate.  In other cases, they
basically attempt to generalize a register-only variant of an instruction into a
version with an immediate, when they can identify a corresponding encoding with
an immediate.

This correspondence is not as simple in non-X86 architectures, so we have to
approach the problem differently.  The core issue is that, once we have a
learned program that behaves the same on all test inputs as our target
instruction, we have to extract a formula representing the target instruction.
Part of that process involves finding the "holes" in the formula to substitute
out as parameters (in a `ParameterizedFormula`).  This is simple for register
operands: they correspond to a `Location` that we can look up directly in a
`Formula` to find a definition.  Immediates are harder, as we can't tell which
term in a formula corresponds to an immediate.  In fact, the immediate may not
appear at all (directly) if it has been simplified out somehow (e.g., through
constant propagation).

Our solution is to lie and cheat.  We pretend that opcodes with immediates
actually only have register operands.  For each immediate operand `I[n]` in an
opcode `O`, we designate a location in the machine state to represent it `L[n]`.
When we need to figure out where an immediate occurs in the formula learned for
a target instruction, we know that it is `L[n]`, since we chose it ourselves.

For example, we might say on PPC that in the `ADDI r1, r2, IMM` instruction,
`IMM` is represented by `r15`.  When we run tests for `ADDI`, we substitute the
value in `r15` in the `ConcreteState` into the `IMM` slot of the instruction.
When we have learned program that implements `ADDI`, we mark the variable
standing in for location `r15` as the parameter corresponding to the immediate.
This lets us generalize learned formulas in a way that preserves immediates.

This approach treats a part of the machine state as an input-only register that
we treat as corresponding to the immediate slot in an instruction.  The learned
program (and formula) has to use that fake immediate slot to have the same
behavior as the target program (which has an immediate instead).  The extra
metadata in `RegisterizedInstruction` is enough to let us pick out which part of
the state to treat in this special way.