Use Stackage releases, not GHC versions in stack.yaml files
We were using GHC versions for resolvers, which meant we needed to
list all dep versions explicitly. Now we use lts for GHC 8.0 and
nightly for GHC 8.2 (there is no lts yet for 8.2), and only list the
package versions that disagree with the stackage release.

Also, lower bound `containers` dep because we use the
`Data.Sequence.lookup` function added in 0.5.8.