Implement test of weighting for right value in wrong place
Stochastic synthesis using "fuzzy matching" for comparing target
output state to candidate output state, where we look for target
values in multiple candidate locations, not just the desired candidate
locations, and then add a small penalty if we find the right value in
the wrong place. Observing random synthesis runs in
`test_synthesizeCandidate`, it appeared this wasn't always happening,
but now I think it was just that we were often getting stuck in a
different local minimum which was worse than producing the right value
in the wrong place (i.e. not producing the right value at all).

In any case, this is a decent integration test.