Simplify handling of `FormulaEnv`
Now the formula loading code takes care of `FormulaEnv` construction itself.
The architecture specific portion (the set of uninterpreted function
descriptions) is now factored out as a method on the `Architecture` class.  It
needs to be in `Architecture` (vs `ConcreteArchitecture`) since it is required
for `TemplatedArch`, which is not an instance of `ConcreteArchitecture`.