Add a helper to the logger to allow for graceful shutdown
Now users of the loggers (the code that invokes a log consumer) can run the
consumer in another thread *and* have the ability to cleanly shut it down (by
calling 'logEndWith').  This is useful to ensure that the logger has finished
processing all of its log messages and successfully flushed them before
terminating the application.