Add stratified synth test to test suite
I had been worried about adding this one, because it's non
deterministic and sometimes takes 10s of seconds, but then I realized
the long run time was because I had only run it interpreted. Compiled
it consistently finishes in under 5 seconds.

As part of this I made the data dir for `runSynToy` configurable, and
moved the Toy arch semantics under `:/semmc`, since Cabal tests are
run relative the package root, not the repo root.

Next up is converting debug printfs in synth to logger calls.