Fix `logM` and `logIOWith` automatic callstack handling
The `logM` was already broken when I first wrote it, and `logIOWith`
was correct when Tristan wrote it, and then I broke it early today
when I refactored it. The point: loggers that automatically infer the
call stack are not composable, in that they infer a call stack at
their call site, so if you use one to build up another log function,
then that derived log function will infer bogus call sites. Of course,
it's pretty easy to write

    writeLogEvent logCfg level msg

when defining a new logger, so not a big deal, just something to watch
out for.