Add more logging and statistics gathering
The statistics changes aren't complete - the backend still has to write the
collected statistics to a durable external location.

This change includes a new constructor for a solver result: Timeout.  The old
code used to just throw an error when the solver times out.  This is not ideal,
so now we handle timeouts more gracefully.