Add support for thread names in logger
Now there is a map from thread IDs to strings in the log config, and a
new function

    asyncNamed :: LogCfg -> String -> IO a -> IO (Async a)

that forks a thread, using the second argument as a human friendly
name in log messages. The `mkLogCfg` also takes a string argument now,
for the names of the main thread that creates the log cfg.

This all works in my test in `SemMC.Log.Tests`, but in full `semmc`
its not really used yet because I just realized the threads I care
about are forked with `timeout`, and so I need name support for
that.

IIRC `timeout` runs the requested action in the current thread, only
spawning a separate thread to run the kill countdown, so it's easy to
wrap `timeout` by creating a version that brackets the action much the
same way that `asyncNamed` does now.