Comment out a line that stops the crashes
It looks like the problem is that the extra threads terminate early due to
having no work -- they then cancel their respective test runners, which causes
the logging thread to crash, which causes a cascade of failures.

This can probably be fixed with more careful handling of cleanup on test runners.