Fix counterexample generation in classify
The theorem prover returns a partial state (only the required parts to form the
counterexample).  Before we use it, we need to flesh it out to be a full
state (i.e., to have all locations set to something).  If we don't, the remote
runner won't receive a full state and will time out on us.