Add a config option for parallel synthesis
This is a separate axis of parallelism from the number of opcodes we learn in
parallel: this lets us try to synthesize candidate programs in parallel, too.

This commit also refactors LocalEnv creation in the Syn monad so that it can be
reused to spawn more threads with local SSH connections.