Fix a major bug in `classify`
Classify was rejecting all new programs because it was including scratch state
when looking for counterexamples.  It should have only been considering the
locations defined by the target instruction (as declared by IORelations).  The
fix is to project out the necessary bits of both formulas and then do the
equivalence check.