[ppc/core] Many improvements to the semantics DSL
Most of the changes involve the type system, which is a bit more elaborated
now.  The changes make it easier to check for consistency without explicitly
passing around bitvector sizes everywhere.  The types are also now more
expressive and avoid some unsafe matches.

In the process of this change, we also added more consistency checks in
bitvector operations -- especially length consistency checks.  This revealed
many sign/zero extension errors that are now fixed.