[ppc] Base set changes
Change the link variants of instructions to have operands of type calltarget
instead of branchtarget.  There isn't really an operational difference there,
but it matches the data.

Also have the link variants of branches declare the IP as an input (since it is
used to compute the new value of the link register).