[ppc] Major backend updates
This commit includes many formula updates, especially around memory operations.

1) There were some operations with the operands in the wrong orders
2) Some uninterpreted function names needed some prefixes since they were
   arch-specific
3) Specialize read_mem and write_mem based on size -- we need this so that we
   can give the operations types in SimpleBuilder

The other big change was a new Parameter type in the formula loader to let us
talk about defining locations referenced through uninterpreted functions over
formula parameters.  We need this to define sub-parts of memri/memrr/memrix
operands in PowerPC.  That work is partly integrated with the synthesis - we'll
see how that plays out when we have a chance to integrate it.

This commit also ports all of the recent changes to the 64 bit backend to the 32
bit backend.

Also fix printing uninterpreted functions in the formula printer.