Refactor ConcreteArchitecture in preparation for removing operandToLocation
The type of operandToLocation is problematic (especially for the floating point
values in PPC), so I am going to try to get rid of it and replace it with
operandToSemanticView.  This will require moving operandToSemanticView from
ConcreteArchitecture to Architecture, which required splitting up the
ConcreteArchitecture module to make the necessary types available.

This change plumbs that change through to all of the affected modules.