[ppc] Start on the floating point base set
This touched a few things, unfortunately.

1) Get rid of the f8rc and f4rc distinction (with accompanying changes in
   dismantle-ppc); we only need one floating point register type.

2) Remove an unused parameter from the floating point uninterpreted functions

3) Introduce some unfortunate renaming churn

This change started out with an attempt to eliminate 'operandToLocation' in
favor of 'operandToSemanticView'.  The problem was that I wanted to make the
operand width of floating point registers 64 bits (or even 32 bits and 64 bits
if we kept f4rc) to make the floating point uninterpreted functions cleaner.
Unfortunately, changing the width of the floating point registers from 128 made
them not match their backing location, which made it impossible to write an
implementation of 'operandToLocation'

Unfortunately, a few places rely on parameter sizes matching location sizes, so
making that change proved very difficult.  It might be possible, but it would
touch a lot of delicate things.

The approach taken in this commit instead is to make the 128 bit backing
locations of the floating point registers (FPRs) explicit: the semantics for the
floating point operations will extract the necessary parts of registers, perform
the requested operations, and then extend the values back into the 128 bit
backing registers.  This should be okay for Crucible, as it simplifies
extract+concat.

The unfortunate renaming arose out of a need for a lot of grepping on some
constructors that had difficult to grep for names.