Make synthesis fail fast when remote runner returns non-zero
E.g. when it returns non-zero by not existing because its not on your
path. Before, the synthesis system would just deadlock or timeout.

However, there's still a race that means sometimes the remote runner
exception kills the synthesis system directly (i.e. the program
returns non-zero), and sometimes the remote runner exception just
appears in the logs, but the program returns zero.