Audit asyncs/forks, to die on all child exceptions
Attempting to make the default that an uncaught exception in a child
thread kills the parent, vs the child silently dying. The basic idea
is that every fork (now all via a variant of `async`) should either be
`link`ed or `wait`ed on. In some cases we use `replicateConcurrently`
or `waitAnyCancel`, but these also propagate exceptions to the parent.

Later, if we discover that some exceptions are recoverable, we can
anticipate them with handlers. But the old default of silently
ignoring exceptions in children seemed error prone.