Add support for most of the conditional branch opcodes
This support is in the defining DSL.  This required adding new operations (a few
bitwise, others logical) to the formula language and parser.  These changes
support normal conditional branches by some offset and conditional branches to
the LNK and CTR registers.

There are still some changes required higher in the stack to support these
fully, especially in higher-level packages that need to support the new
uninterpreted function and the new bitwise operations.