Replace "extract_dynamic" with "test_bit_dynamic"
We only need the restricted operation of test_bit (extract a single bit whose
index is dynamic).  It seems like it would be impossible to write the type of
extract_dynamic in our uninterpreted function framework, as the return value
would be of unknown size.

Use this in the conditional branch defs