[ppc] Fix a bug in the PPC floating point semantics
We have to extract and insert 64 and 32 bit values from the 128 bit register
locations for the non-vector FP instructions.  We were extracting the wrong
part, which meant we were always inspecting the zero-ed out half of the register
instead of the part containing actual values.

This manifest as unused variables, as large chunks of expressions were
eliminated by constant folding and simplification (since we were always
operating on zeros).