[ppc] Add some definitions of some vector registers
This is a first pass covering a few AltiVec and VSX instructions.  There are
many more to go, and many of the definitions are stubs that just set the target
register to undefined.

The loads and stores are more faithfully implemented.  This is plenty for code
discovery for now, but there is still a lot of work to do.