[semmc] Fail fast on empty initial worklist
Down the road we may want this be a success case, but it seems more
likely to indicate a misconfiguration. That's certainly what I'm
dealing with right now: a bunch of manual semantics have been added to
the base set (good), but that means the hard coded target set became a
subset of the base set, which means synthesis was a no-op.