[semmc] Fix stats DB bug
The opcode row had an explicit primary key, but the code that inserted
into the opcode row didn't provide a value for the primary key. So,
instead, we now use the implicit SQLite primary key, i.e. `rowid`.

Also, delete the stats file on start up, to avoid any concerns about
DB migration. In the future we may want to preserve the stats file.