[semmc] Consider changes to implicit operands when weighing candidates
I haven't really tested this yet, since there are no implicit operands
in the Toy arch, and it takes hours to run STRATA on PPC. So, I'm
going to add implicits to the Toy arch.

Also, just noticed the `semmc-toy` candidate weigting test is failing,
but that is actually due to me breaking it in 7c5c96c a little while
back. The problem there is that part of a refactor I made the test
consider some all-zero states for which the right value in the wrong
place is indistinguishable from the right value in the right place,
since all values are zero.