[semmc-toy] Fix right-value-wrong-place test
I broke this test when I added `withInitialState` in 7c5c96c. Before,
we were relying on luck in the right-value-wrong place test: we
assumed that the randomly generated test input states would have
enough 1 bits that the wrong-place penalty would be smaller than the
error in the right place. This is not true for corner cases, which are
unlikely to hit randomly, but when I added `withInitialState` I
refactored the test code to use it, and `withInitialState` includes
the the heuristically interesting tests (read: "corner cases"), where
the old test code did not.

So, I added a bitwise complement instruction to the Toy arch in the
last commit, and use it here to ensure that the value is the right
place is as wrong as possible :)