Improve the formula printer
Change the name to more accurately reflect that it prints parameterized
formulas.  Also extend the printer to support NonceApps, which we use for
uninterpreted functions.  Note that we have a special case for printing
`undefined`, as it has to encode its size in its argument list, but where that
size is not an explicit argument in our uninterpreted function
representation (it is encoded in the return type of the uninterpreted function).