[semmc] Determine when right-value-wrong-place optimization applies
The next step would be to define an arch-indexed interface to `mov` or
`swap` (pseudo) instructions, and then use this interface to implement
moving the wrong place values to the right places. However, the code
in this commit is already enough to tell when and how the optimization
would apply, and hence enough to determine how much implementing the
rest of the optimization would buy us.

This commit also breaks the `semmc-toy` tests, by making STOKE test
fail when the rwvp optimization would apply, but I'm planning to
change that code to log to the stats db instead.