Fix some deprecation warnings from parameterized-utils
This mainly required removing uses of withKnownNat, which is now deprecated.  To
do this, we mostly have to eliminate many KnownNat constraints and introduce
some new ways to work with W (the length-indexed word type).  This should
actually simplify some code in the long run, requiring fewer pattern matches
over the View constructor, as we no longer need to recover KnownNat dictionaries.