[semmc] Log rvwp stats
So that I can compute what effect the optimization would have on
learning speed. Using a the stats collected from learning 1000
candidates for the Toy arch `AddRr` instruction using a two
instruction base set of `SubRr` and `NegR`, and candidate programs of
length 8, I find that rvwp optimization provides a speedup of more
than 2x vs without the optimization, in terms of the number of STOKE
evolution rounds needed to find a candidate on average:

    | num threads | cost w/o opt | cost w/ opt | speedup |
    |-------------+--------------+-------------+---------|
    | 1           | 281          | 133         | 2.11    |
    | 2           | 307          | 169         | 1.81    |
    | 3           | 331          | 200         | 1.65    |
    | 4           | 359          | 217         | 1.65    |
    | 5           | 374          | 246         | 1.52    |

Also interesting, note that the cost per candidate gets *worse* the
more threads we use. Here the cost is the number of rounds needed to
find a candidate. When using multiple threads, the threads race until
one finds a candidate, at which point the others are killed. So, it
turns out the work lost this way doesn't make up for the
outliers (candidate searches that take a long time) that it
eliminates.

Of course, this Toy arch test could produce very different results
from what we'd get on a real arch, since it's very different. Compared
to a real arch this Toy arch config has

- very few instructions in the base set

- no mov or xor instructions in the base set that can effect the mov
  when the rvwp opt applies