[arm] Improve SemARM DSL pred/cond evaluation
The Pred operand is now more directly compared with the appropriate
condition bits in the CPSR (matching the "ConditionPassed" pseudo-code
in the ARM documentation), and the expression to determine this is
stored in the architecture-specific portion of the DSL so that it can
easily be used throughout the remainder of the opcode definition.

Updated "defReg" and "defMem" DSL operations to set the new register
or memory value, based on the ConditionPassed expression.

Resolved some special case handling for the STR_IMM_PRE A32 instruction.