[arm] Use parameterized-utils Assignment to handle opcode arguments.
Opcode arguments are now expressed as an Assignment type instead of
using individual 'param' DSL statements (although the Assignment type
can be a prefix of the actual parameters and the 'param' DSL
statements can still be used for the remainder).

The advantage of this is that the ordering of the arguments can be
consistently defined relative to the Dismantle operands, and that any
helper functions can utilize an argument knowing it has been defined
in the right order (e.g. pred handling for 'defineA32Opcode').