[semmc] Surface more diagnostics when checking for rvwp optimization
When running the PPC stratify it found a candidate where the rvwp
applied with *2* right values in the wrong places. This case is not
implemented, and I didn't think it would ever happen, so now I need to
understand why it's happening. However, I'm optimistic that the wrong
places will be in different "register classes", and so I'll be handle
them by independent movs, and still not need the xor swapping that I'd
use if there were multiple wrong places in the same "register class".