SemMC.TH: expose IO API for loading semantics files
This change refactors the template haskell implementation of semantics
loading in terms of an IO API that is also exported. This enables other
tools to get the semantics contents using the same directory-searching
logic as TH splice users in cases where we don't want to rely on
compiled-in semantics data. (See semmc-coverage for a concrete use of
this API.)