semmc-arm: add A32 architecture
This change adds a new architecture that is specific to the A32
instruction set. This architecture does not include any Thumb
instructions; for that, use the "ARM" architecture. This change also
generalized the uninterpreted functions so they can be used by both the
A32 and ARM architectures: each arch uses different operand/opcode types
and we needed these UFs to be usable for both to avoid duplication.