semmc-arm: generalize some Eval functions over architecture location type
This change generalized some functions in the Eval module so that
they did not assume that they'd be working in terms of the AArch32
location type. This was necessary to re-use these functions with the
A32 architecture that is currently under development, which has many
locations in common but uses its own location type.