arm: Fix a bug in A32 formula evaluators
There are a few accessor functions that return the "add" bit encoded in
operands.  The previous implementation just returned the bits, but the semantics
expected expressions of BaseBoolType.  The fix is to abstract the bitvector
extractor to allow it to wrap results into another term and use it to wrap the bits.