WIP: Fixing a problem with template instantiation
The instantiation of formulas in the synthesis engine currently fails because
there are no evaluators for functions inside of templated formulas.  This branch
is an attempt to re-use the evaluators for the non-templated architectures.

It seems like it is probably possible, but will require a bit of shuffling of
some details.