WIP
Things are now far enough that the expression evaluator can be re-used by the
TemplatedArch without any traumatic things.  The necessary changes actually
simplified the evaluation considerably (including removing the need for the
extra MapF returned by evaluators).

Next is to make sure that the LocationInterpretation function can likewise be
shared.  It looks like it should be possible.