avoid uninterpreted functions in testBitDynamic
There are primitives for checking a bit of a bitvector when the index of
that bit is statically chosen, but none for checking an index that may
be the result of some computation. Previously, testBitDynamic32 and
testBitDynamic64 worked around this problem by creating uninterpreted
functions for that operation, then manually producing an interpretation
in some appropriate places. Unfortunately, it was not enough places: the
concrete semantics evaluator gets stuck at uninterpreted functions.

This commit changes testBitDynamic32 and testBitDynamic64 to use a shift
(which allows dynamically-chosen shift amounts) followed by a testBit on
a static index. Besides enabling the concrete evaluator to make progress
(at least when the arguments are known at runtime), this should also be
easier for solvers to reason about.