Rewrite formula instantiation to fully support the synthesis engine
This change was driven by the so-called "Templated" architecture, which is the
basis of the machine code synthesis engine (MCSE).  The MCSE uses a wrapper type
around the machine architecture to generate instruction /templates/ that are
filled in by a solver.  We try to re-use as much as possible in this code from
the underlying architecture; this became an issue with the
locationFunctionInterpretation, which it turned out was not possible to
implement properly before.  Additional background: the
locationFunctionInterpretation(s) tell SemMC (and the MCSE) how to interpret
some of the functions that appear in semantics formulas.  Examples include:

* Predicates over registers
* Extractors that pull fields out of compound operands (like memory references
  that are composed of a base (register) and an immediate offset)

It happened to be the case that these were implementable for concrete
architectures before (almost as a happy accident), but it was impossible to
re-use those implementations for the Templated architecture, as there were no
longer any concrete operands to inspect to evaluate these functions.  The
Templated architecture has mostly symbolic operands, and our old method for
setting up semantics formulas threw away most of the information we would need
to easily evaluate the location functions.

This change introduces the idea of an AllocatedOperand, which contains all of
the sub-components of operands held separately.  For simple values (literal
values and register/location values) these are simple wrappers around symbolic
values.  For more complicated compound types (like Memrr in PowerPC), we
introduce a facility for architecture-specific data types to hold this
information.  See the various OperandComponents modules.  For example, the
OCMemrr constructor holds:

* The Location (assigned to the register) in the operand
* The symbolic expression allocated to stand in for that location in the solver
* The symbolic expression that represents the offset from the base address of
  the memory reference

Besides making it possible to implement the Templated architecture, keeping the
components separate actually makes evaluating the location functions much
simpler and more efficient (as we don't need to perform as many linear searches
over the operands list anymore).  The way the types work out also makes re-using
evaluators for the ARM architecture easier.

This commit also takes the opportunity to simplify and clean up a few related
issues:
* Remove the RecoverOperandFn type alias and rename WrappedRecoverOperandFn
* Significantly simplify the evaluators in PowerPC
* Add some documentation and choose some better names for a few class methods
* Simplify the type of function Evaluators to not need to return a Map of
  newly-allocated variables anymore; the new approach means that no evaluator
  ever needs to allocate new variables.  Instead, we can always re-use the
  location variables pre-allocated in the formula instantiation code
* Fixed a few incorrect cases in the ARM evaluators
* Clean up the template instantiation code across all architectures (the old
  method was probably never right, and also had to re-produce logic that should
  be entirely encoded in the semantics e.g., adding offsets to base addresses in
  memory references)

Finally, this commit also adds a test suite (with a single test case for now)
that runs the machine code synthesis engine for PowerPC32 and verifies that the
result has the expected semantics.