Fix `MTOCRF` and `MFOCRF`.
The PPC instruction manual states that the field `FXM` contains a 8-bit value
with exactly one bit set to 1, and the index of that bit is the index of the
`CR` field handled by the instruction. However, in Dismantle, the `FXM` field
contains the index directly. Change the SemMC semantics to agree with
Dismantle.