ppc: Introduce a type safe DSL variant for semantics definition
This updated DSL is modeled after the ARM one.  The main distinction is that it
takes an opcode instead of a string for the opcode definition.  The new
definition combinator (definePPCOpcode) enforces that the argument list of the
definition is congruent with the type of the opcode.  The benefit is
that any definition generated will be able to be parsed (since the operand lists
are always guaranteed to be correct).

We also won't be able to write semantics for non-existent opcodes.

Only a few definitions are ported over right now - it can be done incrementally.

Note that the new method fixes the order of operands being passed to the
definition callback, so the order of operands to the lambda are (sometimes)
different from those of the old form combinators (e.g., xoform3 vs xoform3c).

The names can all be fixed up once the porting process is complete.