Handle if statements, add hooks for overrides
The if statement handling uses a type assertion combinator that throws an exception when an
expression doesn't have the expected type.

The hooks allow callers to override the translation of arbitrary expressions or statements.  This
will be useful for handling writes to globals, which are written as function calls (which write to
arrays) in the ASL.