Draft another alternative ProcedureSignature
Always returning the same type (a full register state) makes things a lot simpler (probably).  Also,
we probably needed to do that anyway, given that it isn't always statically known which registers
are written to (without concrete arguments at least).

This version doesn't compile, as the changes need to be plumbed through the rest of the way.