Implement a primitive to capture the register state as a base struct type
There is no way to do this with Crucible syntax, so we have to use an extension.
This requires adding an extra type parameter (arch) to thread through the
arch-specific repr for the register file.

The intent of this is to be able to snapshot the register state before each
procedure call to pass it in.

We'll also need something to put the state back after each call.