Generalize the type of the ASL statement extensions
Instead of always capturing all globals, allow the extension to capture an
arbitrary set of globals.  This will be useful for functions, which typically
only access a few globals.  Eventually, we'll also want to parameterize the
globals for procedure calls in cases where we can recognize statically that they
don't access all globals.