Revise the procedure/function calling convention again
Instead of always passing in the whole register state, pass in only what we
need (which is stored in the signatures of functions and procedures).  This
exposed a few assumptions that were made all over the place, requiring some
tweaks.

This does end up requiring some dynamic checks at call sites to collect just the
set of globals required for a particular call.  This is currently partial.  It
will be very hard to make it total via the type system.