SemMC.Formula.Printer: preserve let-variable binding order
The assignment of let-variables is done in a depth-first
post-order manner over the DAG of expressions. If we maintain
this ordering as we emit let-bindings, it will be easier to
parse the s-expressions later, as the order will coincide
with their dependencies and we can simply parse them in order
while tracking their bindings for subsequent expressions.