Synthesis split (#32)
Split the synthesis libraries from semmc

These libraries aren't related to the core functionality of semmc and are being split out for clarity (making it easier for people learning the library to identify the important components).  Fixes #19. In later commits, we can re-arrange some of the typeclass definitions so that libraries supporting synthesis do not need to directly depend on the implementations of the synthesis algorithms (i.e., by moving the necessary typeclasses into core).