Adapt to GaloisInc/crucible#998
This bumps the `crucible` submodule to bring in the changes from
GaloisInc/crucible#998, which adds `?memOpts :: MemOptions` constraints to
a handful of additional functions. This requires adding constraints to
some functions in `semmc-synthesis` to accommodate.

The `crucible-llvm` submodule now depends on a more recent `llvm-pretty`
submodule commit to build, so this patch also bumps the `llvm-pretty` submodule.