Bump what4 submodule to version 1.4
This version of `what4` includes the changes from GaloisInc/what4#217, which
merges the `what4-serialize` library into `what4`. As a result, some of the
imports in `semmc` had to be adjusted somewhat.

I also bumped the `asl-translator` submodule to bring in the changes from
GaloisInc/asl-translator#48, which performed a similar `what4` adaptation. I
also bumped the `crucible` submodule to bring in the changes from
GaloisIns/crucible#1068, which ensures that everything can build against
`tasty-sugar >= 2.0` (the version of the library that `what4-1.4` depends on).